Выделил отдельно popup presenter и переделал логику отрисовки изменений
