+++ src/presenter/film.js
@@ -1,13 +1,16 @@
+import AbstractFilmPresenter from './abstract-film.js';
 import FilmCardBlockView from '../view/film-card-block.js';
 import { render, remove, replace } from '../utils/render.js';
-import { BUTTON_TYPE } from '../const.js';
+//import { BUTTON_TYPE } from '../const.js';
 
-export default class FilmPresenter {
+export default class FilmPresenter extends AbstractFilmPresenter {
   constructor(container, handleFilmChange) {
+    super();
     this._filmContainer = container;
-    this._changeData = handleFilmChange;
     this._filmComponent = null;
-    this.handleControlButtons = this.handleControlButtons.bind(this);
+    //_changeData и _handleControlButtons наследуются от AbstractFilmPresenter
+    this._changeData = handleFilmChange;
+    this._handleControlButtons = this._handleControlButtons.bind(this);
   }
 
   init(film) {
@@ -15,7 +18,7 @@ export default class FilmPresenter {
     const oldFilmComponent = this._filmComponent;
 
     this._filmComponent = new FilmCardBlockView(this._film);
-    this._filmComponent.setControlButtonsClick(this.handleControlButtons);
+    this._filmComponent.setControlButtonsClick(this._handleControlButtons);
 
     if (oldFilmComponent === null) {
       render(this._filmContainer, this._filmComponent);
@@ -30,29 +33,29 @@ export default class FilmPresenter {
     remove(this._filmComponent);
   }
 
-  handleControlButtons(evt) {
-    const buttonType = evt.target.dataset.type;
-    // делаем копию данных фильма
-    const changedUserDetails = {...this._film.userDetails};
+  // handleControlButtons(evt) {
+  //   const buttonType = evt.target.dataset.type;
+  //   // делаем копию данных фильма
+  //   const changedUserDetails = {...this._film.userDetails};
 
-    if (!buttonType) {
-      return;
-    }
-    // изменяем ключи на противоположное значение
-    // в зависимости от типа кнопки
-    switch (buttonType) {
-      case BUTTON_TYPE.watchlisted:
-        changedUserDetails.isWatchlisted = !this._film.userDetails.isWatchlisted;
-        break;
-      case BUTTON_TYPE.watched:
-        changedUserDetails.isWatched = !this._film.userDetails.isWatched;
-        break;
-      case BUTTON_TYPE.favorite:
-        changedUserDetails.isFavorite = !this._film.userDetails.isFavorite;
-        break;
-    }
+  //   if (!buttonType) {
+  //     return;
+  //   }
+  //   // изменяем ключи на противоположное значение
+  //   // в зависимости от типа кнопки
+  //   switch (buttonType) {
+  //     case BUTTON_TYPE.watchlisted:
+  //       changedUserDetails.isWatchlisted = !this._film.userDetails.isWatchlisted;
+  //       break;
+  //     case BUTTON_TYPE.watched:
+  //       changedUserDetails.isWatched = !this._film.userDetails.isWatched;
+  //       break;
+  //     case BUTTON_TYPE.favorite:
+  //       changedUserDetails.isFavorite = !this._film.userDetails.isFavorite;
+  //       break;
+  //   }
 
-    // и передаем объект с измененными данными фильма
-    this._changeData({...this._film, userDetails: changedUserDetails});
-  }
+  //   // и передаем объект с измененными данными фильма
+  //   this._changeData({...this._film, userDetails: changedUserDetails});
+  // }
 }
