+++ src/presenter/films-list.js
@@ -5,9 +5,9 @@ import FilmsByRatingView from '../view/films-by-rating.js';
 import FilmsByCommentsView from '../view/films-by-comments.js';
 import NoFilmsBlockView from '../view/no-films-block.js';
 import ShowMoreButtonView from '../view/button-show-more.js';
-import FilmPopupView from '../view/film-popup.js';
 import FilmPresenter from '../presenter/film.js';
-import { render, remove, replace } from '../utils/render.js';
+import PopupPresenter from '../presenter/film-popup.js';
+import { render, remove } from '../utils/render.js';
 import { updateItem, sortByDate, sortByRating } from '../utils/common.js';
 import { FILMS_PER_STEP, SORT_BY, SORT_TYPE, EXTRA_FILMS_CARDS_COUNT } from '../const.js';
 
@@ -26,14 +26,13 @@ export default class FilmsList {
     this._filmsByCommentsComponent = new FilmsByCommentsView();
     this._noFilmsBlockComponent = new NoFilmsBlockView();
     this._buttonShowMoreComponent = new ShowMoreButtonView();
-    this._filmPopupComponent = null;
+    this._popupPresenter = null;
 
     this._handleButtonShowMore = this._handleButtonShowMore.bind(this);
     this._handleFilmsList = this._handleFilmsList.bind(this);
-    this._handleClosePopupButton = this._handleClosePopupButton.bind(this);
-    this._handleEscKeyDown = this._handleEscKeyDown.bind(this);
     this._handleFilmChange = this._handleFilmChange.bind(this);
     this._handleSortButtons = this._handleSortButtons.bind(this);
+    this._clearPopupPresenter = this._clearPopupPresenter.bind(this);
 
     this._filmsListSection = this._filmsSectionComponent.getElement().querySelector('.films-list');
     this._filmsListContainer = this._filmsListSection.querySelector('.films-list__container');
@@ -53,8 +52,6 @@ export default class FilmsList {
       this._renderFilmsContainer();
       this._renderNoFilmsBlock();
     }
-    //console.log(this._filmPresentersList);
-    //console.log(this._filmPresentersList.values());
   }
 
   //=====
@@ -130,45 +127,15 @@ export default class FilmsList {
     this._buttonShowMoreComponent.setClickHandler(this._handleButtonShowMore);
   }
 
-  _renderPopup(film) {
-    const oldPopupComponent = this._filmPopupComponent;
-    this._filmPopupComponent = new FilmPopupView(film, this._comments);
-    // рендерим попап
-    if (oldPopupComponent === null) {
-      render(this._filmsContainer, this._filmPopupComponent);
-    } else {
-      // если попап открыт перерисовываем
-      const scrollY = oldPopupComponent.getElement().scrollTop;
-      replace(oldPopupComponent, this._filmPopupComponent);
-      this._filmPopupComponent.getElement().scrollTop = scrollY;
-      remove(oldPopupComponent);
-    }
-    // находим экземпляр презентера фильма
-    let filmPresenter;
-    this._filmPresentersList.forEach((id, presenter) => {
-      if (id === film.filmInfo.id) {
-        filmPresenter = presenter;
-      }
-    });
-    // навешиваем нужные классы и слушатели
-    document.body.classList.add('hide-overflow');
-    document.addEventListener('keydown', this._handleEscKeyDown);
+  _renderPopup(container, film, callback) {
+    this._popupPresenter = new PopupPresenter(container, this._handleFilmChange, callback);
+    this._popupPresenter.init(film, this._comments);
     this._filmsSectionComponent.removeFilmCardClickHandler();
-    this._filmPopupComponent.setCloseButtonClickHandler(this._handleClosePopupButton);
-    this._filmPopupComponent.setControlButtonsClick(filmPresenter.handleControlButtons);
   }
 
   //=====
-  // Методы удаления, сортировки и очистки
+  // Методы очистки и сортировки
   //=====
-
-  _removePopup() {
-    remove(this._filmPopupComponent);
-    this._filmPopupComponent = null;
-    document.body.classList.remove('hide-overflow');
-    document.removeEventListener('keydown', this._handleEscKeyDown);
-    this._filmsSectionComponent.setFilmCardClickHandler(this._handleFilmsList);
-  }
 
   _clearFilmsList() {
     this._filmPresentersList
@@ -176,6 +143,11 @@ export default class FilmsList {
     this._filmPresentersList.clear();
     this._renderedFilmsCount = FILMS_PER_STEP;
     remove(this._buttonShowMoreComponent);
+  }
+
+  _clearPopupPresenter() {
+    this._popupPresenter = null;
+    this._filmsSectionComponent.setFilmCardClickHandler(this._handleFilmsList);
   }
 
   _sortFilms(sortType) {
@@ -221,19 +193,7 @@ export default class FilmsList {
     const filmCardId = target.closest('.film-card').dataset.id;
     const film = this._films.find(({filmInfo}) => filmCardId === filmInfo.id);
     // рендерим попап
-    this._renderPopup(film);
-  }
-
-  // обработчик закрытия попапа
-  _handleClosePopupButton() {
-    this._removePopup();
-  }
-
-  // обработчик Esc
-  _handleEscKeyDown(evt) {
-    if (evt.key === 'Escape' || evt.key === 'Esc') {
-      this._removePopup();
-    }
+    this._renderPopup(this._filmsContainer, film, this._clearPopupPresenter);
   }
 
   // обработчик изменения данных фильма
@@ -249,9 +209,9 @@ export default class FilmsList {
         component.init(updatedFilm);
       }
     });
-    // перерисовка попапа, если есть
-    if (this._filmPopupComponent !== null) {
-      this._renderPopup(updatedFilm);
+    // и перерисовку попапа, если он открыт
+    if (this._popupPresenter !== null) {
+      this._popupPresenter.init(updatedFilm, this._comments);
     }
   }
